+++ Trabalho de SO/io.h
@@ -1,0 +1,20 @@
+//
+//  io.h
+//  Trabalho de SO
+//
+//  Created by Rafael Lopes Francisco on 30/05/15.
+//  Copyright (c) 2015 Rafael Lopes Francisco. All rights reserved.
+//
+
+#ifndef __Trabalho_de_SO__io__
+#define __Trabalho_de_SO__io__
+#define TAM 256
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include "structs_defs.h"
+
+sala * carregarDados();
+
+#endif /* defined(__Trabalho_de_SO__io__) */
